+++ src/main.rs
@@ -27,6 +27,7 @@ fn main() {
         }
     };
     let mut vm = VMRef::new(VM::new());
+    vm.clone().globals.fibers.push(vm);
     let id = vm.globals.get_ident_id("ARGV");
     let mut res: Vec<Value> = args
         .iter()
