+++ src/repl.rs
@@ -22,7 +22,8 @@ pub fn repl_vm() {
     let mut rl = rustyline::Editor::<()>::new();
     let mut program = String::new();
     let mut parser = Parser::new();
-    let mut vm = VM::new();
+    let mut vm = VMRef::new(VM::new());
+    vm.clone().globals.fibers.push(vm);
     parser.ident_table = vm.globals.ident_table.clone();
     let mut level = parser.get_context_depth();
     let mut lvar_collector = LvarCollector::new();
