+++ src/test.rs
@@ -2,7 +2,8 @@ use crate::*;
 use std::path::PathBuf;
 
 pub fn eval_script(script: impl Into<String>, expected: Value) {
-    let mut vm = VM::new();
+    let mut vm = VMRef::new(VM::new());
+    vm.clone().globals.fibers.push(vm);
     match vm.run(PathBuf::from(""), &script.into(), None) {
         Ok(res) => {
             if res != expected {
@@ -18,7 +19,8 @@ pub fn eval_script(script: impl Into<String>, expected: Value) {
 }
 
 pub fn assert_script(script: impl Into<String>) {
-    let mut vm = VM::new();
+    let mut vm = VMRef::new(VM::new());
+    vm.clone().globals.fibers.push(vm);
     match vm.run(PathBuf::from(""), &script.into(), None) {
         Ok(_) => {}
         Err(err) => {
